Fix 5 bugs found in full codebase audit
1. ble_notifications.h: Split into .h (declarations only) + .cpp
   (implementations). Having globals and function bodies in the header
   caused multiple-definition linker errors since notification_service.cpp
   and SmartWatchV3.ino both included it as separate translation units.

2. Alert server never started: ConnectivityService::startAlertServer()
   was implemented but never called. Now called when WiFi connects
   via the control panel toggle. Also stops server on WiFi disconnect.

3. Focus timer overflow: _totalDurationUs was uint32_t, max ~71 minutes.
   Changed to int64_t so timers >71 minutes work correctly.

4. Control panel BLE status hardcoded "Connected": Now reads actual
   connection state via isBLEConnected().

5. Notification swipe-dismiss race: Rebuilding the list was delayed
   200ms, allowing a second dismiss to use a stale index. Now rebuilds
   immediately after dismiss.

+++ SmartWatchV3/src/services/focus_service.cpp
@@ -11,7 +11,7 @@
 #include <Arduino.h>
 
 static FocusState _state = FOCUS_IDLE;
-static uint32_t _totalDurationUs = 0;     // total in microseconds
+static int64_t _totalDurationUs = 0;      // total in microseconds (int64 to support >71 min)
 static int64_t _startTimeUs = 0;          // when timer started (absolute)
 static int64_t _pausedElapsedUs = 0;      // elapsed before pause
 
@@ -23,7 +23,7 @@ void FocusService::init(void) {
 }
 
 void FocusService::startTimer(uint16_t minutes) {
-    _totalDurationUs = (uint32_t)minutes * 60UL * 1000000UL;
+    _totalDurationUs = (int64_t)minutes * 60LL * 1000000LL;
     _startTimeUs = esp_timer_get_time();
     _pausedElapsedUs = 0;
     _state = FOCUS_ACTIVE;
@@ -67,7 +67,7 @@ uint32_t FocusService::getRemainingSeconds(void) {
         elapsedUs = _pausedElapsedUs + (esp_timer_get_time() - _startTimeUs);
     }
 
-    if (elapsedUs >= (int64_t)_totalDurationUs) return 0;
+    if (elapsedUs >= _totalDurationUs) return 0;
 
     return (uint32_t)((_totalDurationUs - elapsedUs) / 1000000UL);
 }
@@ -99,7 +99,7 @@ void FocusService::update(void) {
     if (_state != FOCUS_ACTIVE) return;
 
     int64_t elapsedUs = _pausedElapsedUs + (esp_timer_get_time() - _startTimeUs);
-    if (elapsedUs >= (int64_t)_totalDurationUs) {
+    if (elapsedUs >= _totalDurationUs) {
         _state = FOCUS_COMPLETE;
         Serial.println("Focus: timer complete!");
     }
